Remove mail-server link from admin tab and Remove Mail-server SPA as it is merged in Server -config SPA.
