+++ server/src/main/webapp/WEB-INF/rails/webpack/views/components/site_menu/index.tsx
@@ -167,7 +167,6 @@ export class SiteMenu extends MithrilViewComponent<Attrs> {
               <SiteSubNav>
                 <SiteSubNavHeading text="Server configuration"/>
                 <SiteSubNavItem href="/go/admin/config/server" text="Server Configuration"/>
-                <SiteSubNavItem href="/go/admin/mail_server" text="Email server"/>
                 <SiteSubNavItem href="/go/admin/data_sharing/settings" text="Data Sharing"/>
                 <SiteSubNavItem href="/go/admin/maintenance_mode" text="Server Maintenance Mode"/>
                 <SiteSubNavItem href="/go/admin/backup" text="Backup"/>
