--- server/src/main/webapp/WEB-INF/rails/webpack/views/pages/mail_server/index.scss.d.ts
@@ -1,3 +1,0 @@
-// This file is automatically generated. Please do not change this file!
-export const content: string;
-export const mailServerContainer: string;
